Add badge to app icon
Migrate to using an ApplicationDelegate in order to run code on
start, as opposed to running it on the app's init (which could be
called many times).

Let the delegate keep track of updating the feeds.

Add a badge to the icon by using ImageRenderer, which renders a
view any time the feeds are changed. Makes it more clear that there
are updates to be read.